+++ js/script.js
@@ -1,3 +1,4 @@
+
 const allCountriesPhonCodes =
 [ {dial_code: "+1" , value:"us", country: "United States"}, {dial_code: "+93" , value:"af", country: "Afghanistan"}, {dial_code: "+355" , value:"al", country: "Albania"}, {dial_code: "+213" , value:"dz", country: "Algeria"}, {dial_code: "+1684" , value:"as", country: "American Samoa"}, {dial_code: "+376" , value:"ad", country: "Andorra"}, {dial_code: "+244" , value:"ao", country: "Angola"}, {dial_code: "+1264" , value:"ai", country: "Anguilla"}, {dial_code: "+672" , value:"aq", country: "Antarctica"}, {dial_code: "+1" , value:"ag", country: "Antigua and Barbuda"}, {dial_code: "+54" , value:"ar", country: "Argentina"}, {dial_code: "+374" , value:"am", country: "Armenia"}, {dial_code: "+297" , value:"aw", country: "Aruba"}, {dial_code: "+61" , value:"au", country: "Australia"}, {dial_code: "+43" , value:"at", country: "Austria"}, {dial_code: "+994" , value:"az", country: "Azerbaijan"}, {dial_code: "+1242" , value:"bs", country: "Bahamas"}, {dial_code: "+973" , value:"bh", country: "Bahrain"}, {dial_code: "+880" , value:"bd", country: "Bangladesh"}, {dial_code: "+1246" , value:"bb", country: "Barbados"}, {dial_code: "+375" , value:"by", country: "Belarus"}, {dial_code: "+32" , value:"be", country: "Belgium"}, {dial_code: "+501" , value:"bz", country: "Belize"}, {dial_code: "+229" , value:"bj", country: "Benin"}, {dial_code: "+975" , value:"bt", country: "Bhutan"}, {dial_code: "+591" , value:"bo", country: "Bolivia"}, {dial_code: "+387" , value:"ba", country: "Bosnia and Herzegovina"}, {dial_code: "+267" , value:"bw", country: "Botswana"}, {dial_code: "+55" , value:"br", country: "Brazil"}, {dial_code: "+673" , value:"bn", country: "Brunei Darussalam"}, {dial_code: "+359" , value:"bg", country: "Bulgaria"}, {dial_code: "+226" , value:"bf", country: "Burkina Faso"}, {dial_code: "+257" , value:"bi", country: "Burundi"}, {dial_code: "+855" , value:"kh", country: "Cambodia"}, {dial_code: "+237" , value:"cm", country: "Cameroon"}, {dial_code: "+238" , value:"cv", country: "Cape Verde"}, {dial_code: "+236" , value:"cf", country: "Central African Republic"}, {dial_code: "+235" , value:"td", country: "Chad"}, {dial_code: "+56" , value:"cl", country: "Chile"}, {dial_code: "+86" , value:"cn", country: "China"}, {dial_code: "+61" , value:"cx", country: "Christmas Island"}, {dial_code: "+57" , value:"co", country: "Colombia"}, {dial_code: "+269" , value:"km", country: "Comoros"}, {dial_code: "+242" , value:"cg", country: "Congo"}, {dial_code: "+682" , value:"ck", country: "Cook Islands"}, {dial_code: "+506" , value:"cr", country: "Costa Rica"}, {dial_code: "+225" , value:"ci", country: "Cote D’Ivoire (Ivory Coast)"}, {dial_code: "+385" , value:"hr", country: "Croatia"}, {dial_code: "+53" , value:"cu", country: "Cuba"}, {dial_code: "+357" , value:"cy", country: "Cyprus"}, {dial_code: "+420" , value:"cz", country: "Czech Republic"}, {dial_code: "+243" , value:"cd", country: "Democratic Republic of the Congo"}, {dial_code: "+45" , value:"dk", country: "Denmark"}, {dial_code: "+253" , value:"dj", country: "Djibouti"}, {dial_code: "+1767" , value:"dm", country: "Dominica"}, {dial_code: "+1809" , value:"do", country: "Dominican Republic"}, {dial_code: "+593" , value:"ec", country: "Ecuador"}, {dial_code: "+20" , value:"eg", country: "Egypt"}, {dial_code: "+503" , value:"sv", country: "El Salvador"}, {dial_code: "+240" , value:"gq", country: "Equatorial Guinea"}, {dial_code: "+291" , value:"er", country: "Eritrea"}, {dial_code: "+372" , value:"ee", country: "Estonia"}, {dial_code: "+251" , value:"et", country: "Ethiopia"}, {dial_code: "+500" , value:"fk", country: "Falkland Islands (Malvinas)"}, {dial_code: "+298" , value:"fo", country: "Faroe Islands"}, {dial_code: "+691" , value:"fm", country: "Federated States of Micronesia"}, {dial_code: "+679" , value:"fj", country: "Fiji"}, {dial_code: "+358" , value:"fi", country: "Finland"}, {dial_code: "+33" , value:"fr", country: "France"}, {dial_code: "+594" , value:"gf", country: "French Guiana"}, {dial_code: "+689" , value:"pf", country: "French Polynesia"}, {dial_code: "+241" , value:"ga", country: "Gabon"}, {dial_code: "+220" , value:"gm", country: "Gambia"}, {dial_code: "+995" , value:"ge", country: "Georgia"}, {dial_code: "+49" , value:"de", country: "Germany"}, {dial_code: "+233" , value:"gh", country: "Ghana"}, {dial_code: "+350" , value:"gi", country: "Gibraltar"}, {dial_code: "+30" , value:"gr", country: "Greece"}, {dial_code: "+299" , value:"gl", country: "Greenland"}, {dial_code: "+1473" , value:"gd", country: "Grenada"}, {dial_code: "+590" , value:"gp", country: "Guadeloupe"}, {dial_code: "+1671" , value:"gu", country: "Guam"}, {dial_code: "+502" , value:"gt", country: "Guatemala"}, {dial_code: "+224" , value:"gn", country: "Guinea"}, {dial_code: "+245" , value:"gw", country: "Guinea-Bissau"}, {dial_code: "+592" , value:"gy", country: "Guyana"}, {dial_code: "+509" , value:"ht", country: "Haiti"}, {dial_code: "+504" , value:"hn", country: "Honduras"}, {dial_code: "+852" , value:"hk", country: "Hong Kong"}, {dial_code: "+36" , value:"hu", country: "Hungary"}, {dial_code: "+354" , value:"is", country: "Iceland"}, {dial_code: "+91" , value:"in", country: "India"}, {dial_code: "+62" , value:"id", country: "Indonesia"}, {dial_code: "+98" , value:"ir", country: "Iran"}, {dial_code: "+964" , value:"iq", country: "Iraq"}, {dial_code: "+353" , value:"ie", country: "Ireland"}, {dial_code: "+972" , value:"il", country: "Israel"}, {dial_code: "+39" , value:"it", country: "Italy"}, {dial_code: "+1" , value:"jm", country: "Jamaica"}, {dial_code: "+81" , value:"jp", country: "Japan"}, {dial_code: "+962" , value:"jo", country: "Jordan"}, {dial_code: "+7" , value:"kz", country: "Kazakhstan"}, {dial_code: "+254" , value:"ke", country: "Kenya"}, {dial_code: "+686" , value:"ki", country: "Kiribati"}, {dial_code: "+82" , value:"kr", country: "Korea"}, {dial_code: "+850" , value:"kp", country: "Korea (North)"}, {dial_code: "+965" , value:"kw", country: "Kuwait"}, {dial_code: "+996" , value:"kg", country: "Kyrgyzstan"}, {dial_code: "+856" , value:"la", country: "Laos"}, {dial_code: "+371" , value:"lv", country: "Latvia"}, {dial_code: "+961" , value:"lb", country: "Lebanon"}, {dial_code: "+266" , value:"ls", country: "Lesotho"}, {dial_code: "+231" , value:"lr", country: "Liberia"}, {dial_code: "+218" , value:"ly", country: "Libya"}, {dial_code: "+423" , value:"li", country: "Liechtenstein"}, {dial_code: "+370" , value:"lt", country: "Lithuania"}, {dial_code: "+352" , value:"lu", country: "Luxembourg"}, {dial_code: "+853" , value:"mo", country: "Macao"}, {dial_code: "+389" , value:"mk", country: "Macedonia"}, {dial_code: "+261" , value:"mg", country: "Madagascar"}, {dial_code: "+265" , value:"mw", country: "Malawi"}, {dial_code: "+60" , value:"my", country: "Malaysia"}, {dial_code: "+960" , value:"mv", country: "Maldives"}, {dial_code: "+223" , value:"ml", country: "Mali"}, {dial_code: "+356" , value:"mt", country: "Malta"}, {dial_code: "+692" , value:"mh", country: "Marshall Islands"}, {dial_code: "+596" , value:"mq", country: "Martinique"}, {dial_code: "+222" , value:"mr", country: "Mauritania"}, {dial_code: "+230" , value:"mu", country: "Mauritius"}, {dial_code: "+269" , value:"yt", country: "Mayotte"}, {dial_code: "+52" , value:"mx", country: "Mexico"}, {dial_code: "+373" , value:"md", country: "Moldova"}, {dial_code: "+377" , value:"mc", country: "Monaco"}, {dial_code: "+976" , value:"mn", country: "Mongolia"}, {dial_code: "+382" , value:"me", country: "Montenegro"}, {dial_code: "+1664" , value:"ms", country: "Montserrat"}, {dial_code: "+212" , value:"ma", country: "Morocco"}, {dial_code: "+258" , value:"mz", country: "Mozambique"}, {dial_code: "+95" , value:"mm", country: "Myanmar"}, {dial_code: "+264" , value:"na", country: "Namibia"}, {dial_code: "+674" , value:"nr", country: "Nauru"}, {dial_code: "+977" , value:"np", country: "Nepal"}, {dial_code: "+31" , value:"nl", country: "Netherlands"}, {dial_code: "+599" , value:"an", country: "Netherlands Antilles"}, {dial_code: "+687" , value:"nc", country: "New Caledonia"}, {dial_code: "+64" , value:"nz", country: "New Zealand"}, {dial_code: "+505" , value:"ni", country: "Nicaragua"}, {dial_code: "+227" , value:"ne", country: "Niger"}, {dial_code: "+234" , value:"ng", country: "Nigeria"}, {dial_code: "+683" , value:"nu", country: "Niue"}, {dial_code: "+672" , value:"nf", country: "Norfolk Island"}, {dial_code: "+1670" , value:"mp", country: "Northern Mariana Islands"}, {dial_code: "+47" , value:"no", country: "Norway"}, {dial_code: "+92" , value:"pk", country: "Pakistan"}, {dial_code: "+680" , value:"pw", country: "Palau"}, {dial_code: "+970" , value:"ps", country: "Palestinian Territory"}, {dial_code: "+507" , value:"pa", country: "Panama"}, {dial_code: "+675" , value:"pg", country: "Papua New Guinea"}, {dial_code: "+595" , value:"py", country: "Paraguay"}, {dial_code: "+51" , value:"pe", country: "Peru"}, {dial_code: "+63" , value:"ph", country: "Philippines"}, {dial_code: "+870" , value:"pn", country: "Pitcairn"}, {dial_code: "+48" , value:"pl", country: "Poland"}, {dial_code: "+351" , value:"pt", country: "Portugal"}, {dial_code: "+974" , value:"qa", country: "Qatar"}, {dial_code: "+262" , value:"re", country: "Reunion"}, {dial_code: "+40" , value:"ro", country: "Romania"}, {dial_code: "+7" , value:"ru", country: "Russian Federation"}, {dial_code: "+250" , value:"rw", country: "Rwanda"}, {dial_code: "+290" , value:"sh", country: "Saint Helena"}, {dial_code: "+1869" , value:"kn", country: "Saint Kitts and Nevis"}, {dial_code: "+508" , value:"pm", country: "Saint Pierre and Miquelon"}, {dial_code: "+685" , value:"ws", country: "Samoa"}, {dial_code: "+378" , value:"sm", country: "San Marino"}, {dial_code: "+239" , value:"st", country: "Sao Tome and Principe"}, {dial_code: "+966" , value:"sa", country: "Saudi Arabia"}, {dial_code: "+221" , value:"sn", country: "Senegal"}, {dial_code: "+381" , value:"rs", country: "Serbia"}, {dial_code: "+248" , value:"sc", country: "Seychelles"}, {dial_code: "+232" , value:"sl", country: "Sierra Leone"}, {dial_code: "+65" , value:"sg", country: "Singapore"}, {dial_code: "+421" , value:"sk", country: "Slovak Republic"}, {dial_code: "+386" , value:"si", country: "Slovenia"}, {dial_code: "+677" , value:"sb", country: "Solomon Islands"}, {dial_code: "+252" , value:"so", country: "Somalia"}, {dial_code: "+27" , value:"za", country: "South Africa"}, {dial_code: "+34" , value:"es", country: "Spain"}, {dial_code: "+94" , value:"lk", country: "Sri Lanka"}, {dial_code: "+249" , value:"sd", country: "Sudan"}, {dial_code: "+968" , value:"om", country: "Sultanate of Oman"}, {dial_code: "+597" , value:"sr", country: "Suriname"}, {dial_code: "+268" , value:"sz", country: "Swaziland"}, {dial_code: "+46" , value:"se", country: "Sweden"}, {dial_code: "+41" , value:"ch", country: "Switzerland"}, {dial_code: "+963" , value:"sy", country: "Syria"}, {dial_code: "+886" , value:"tw", country: "Taiwan"}, {dial_code: "+992" , value:"tj", country: "Tajikistan"}, {dial_code: "+255" , value:"tz", country: "Tanzania"}, {dial_code: "+66" , value:"th", country: "Thailand"}, {dial_code: "+670" , value:"tl", country: "Timor-Leste"}, {dial_code: "+228" , value:"tg", country: "Togo"}, {dial_code: "+690" , value:"tk", country: "Tokelau"}, {dial_code: "+676" , value:"to", country: "Tonga"}, {dial_code: "+1868" , value:"tt", country: "Trinidad and Tobago"}, {dial_code: "+216" , value:"tn", country: "Tunisia"}, {dial_code: "+90" , value:"tr", country: "Turkey"}, {dial_code: "+993" , value:"tm", country: "Turkmenistan"}, {dial_code: "+1649" , value:"tc", country: "Turks and Caicos Islands"}, {dial_code: "+688" , value:"tv", country: "Tuvalu"}, {dial_code: "+256" , value:"ug", country: "Uganda"}, {dial_code: "+380" , value:"ua", country: "Ukraine"}, {dial_code: "+971" , value:"ae", country: "United Arab Emirates"}, {dial_code: "+44" , value:"gb", country: "United Kingdom"}, {dial_code: "+598" , value:"uy", country: "Uruguay"}, {dial_code: "+998" , value:"uz", country: "Uzbekistan"}, {dial_code: "+678" , value:"vu", country: "Vanuatu"}, {dial_code: "+39" , value:"va", country: "Vatican City State (Holy See)"}, {dial_code: "+58" , value:"ve", country: "Venezuela"}, {dial_code: "+84" , value:"vn", country: "Vietnam"}, {dial_code: "+1284" , value:"vg", country: "Virgin Islands (British)"}, {dial_code: "+1340" , value:"vi", country: "Virgin Islands (U.S.)"}, {dial_code: "+681" , value:"wf", country: "Wallis and Futuna"}, {dial_code: "+967" , value:"ye", country: "Yemen"}, {dial_code: "+260" , value:"zm", country: "Zambia"}, {dial_code: "+263" , value:"zw", country: "Zimbabwe"}, ]
 
@@ -5,6 +6,7 @@ let email
 
 if(window.location.href.includes("linkedin.com/signup")) {
     email = getEmail()
+    // joinFormEditor()
     createEmailPasteButton()
     createButtonsObserver()
 }
@@ -145,6 +147,27 @@ function getRandomData () {
 //     localStorage.setItem("EmailHystory", JSON.stringify(emailHystory))
 // }
 
+// function joinFormEditor() {
+//     if(!window.location.href.includes("linkedin.com/signup")) return null
+//     const form = document.querySelector("#join-form-submit")
+//     // const formListeners = getEventListeners(document.querySelector('.join-form'));
+//     // console.log(formListeners)
+//     form.onclick= async function (e) {
+        
+//             // e.preventDefault()
+//             console.dir(form)
+//             const clipData = await getDataFromClipboard()
+//             email.value = ValidateEmail(clipData)
+//             getRandomData()
+//             console.log("done")
+            
+//         }
+//     console.dir(form)
+    
+        
+    
+// }
+
 function createEmailPasteButton() {
     
     if(!window.location.href.includes("linkedin.com/signup")) return null
@@ -160,6 +183,9 @@ function createEmailPasteButton() {
             const clipData = await getDataFromClipboard()
             email.value = ValidateEmail(clipData)
             getRandomData()
+           let btnSubmit = document.getElementById('join-form-submit')
+           btnSubmit.click()
+           btnSubmit.onload = btnSubmit.click()
         
     })} else return 
 
@@ -169,6 +195,7 @@ function createNumberPasteButtonsGrupe() {
     if(!window.location.href.includes("linkedin.com/signup")) return null
     let buttonsContainer = window.frames[0].document.getElementById('number-buttons-container')
     let number = window.frames[0].document.getElementById('register-verification-phone-number')
+    const registerPhoneSubmitBtn = window.frames[0].document.getElementById('register-phone-submit-button')
     
     if (buttonsContainer === null && number !== null) {
         
@@ -184,6 +211,8 @@ function createNumberPasteButtonsGrupe() {
         const countryInfoByDialCode = getCountryInfoByNumber(clipData)
         number.value = itNumber(countryInfoByDialCode?.fullNumber)
         phoneCountrySelector(countryInfoByDialCode.selectValue)
+        registerPhoneSubmitBtn.click()
+        
         })
         let withoutDialingCodeNumberPasteButton = window.frames[0].document.querySelector("#without-dialing-code-paste-button")
         withoutDialingCodeNumberPasteButton.addEventListener("click", async e => {
@@ -192,6 +221,7 @@ function createNumberPasteButtonsGrupe() {
         const countryInfoByDialCode = getCountryInfoByNumber(clipData)
         number.value = itNumber(countryInfoByDialCode?.numberWithOutDialCode)
         phoneCountrySelector(countryInfoByDialCode.selectValue)
+        registerPhoneSubmitBtn.click()
         })
     } else return   
 }
@@ -200,7 +230,8 @@ function createPinPasteButton() {
     if(!window.location.href.includes("linkedin.com/signup")) return null
     let pinPasteButton = window.frames[0].document.getElementById('pin-paste-button')
     let pin = window.frames[0].document.getElementById('input__phone_verification_pin')
-    
+    const registerPhoneSubmitBtn = window.frames[0].document.getElementById('register-phone-submit-button')
+
     if (pinPasteButton === null && pin !== null) {
         pin.parentNode.after(createContainer("pin"))
         let buttonsContainer = window.frames[0].document.getElementById('pin-buttons-container')
@@ -210,6 +241,7 @@ function createPinPasteButton() {
         e.preventDefault()
         const clipData = await getDataFromClipboard()
         pin.value = clipData
+        registerPhoneSubmitBtn.click()
     })} else return   
 }
 
@@ -300,7 +332,7 @@ function getCountryInfoByNumber(number) {
     {
       return clipData
     }
-      return "there is no email in the clipboard, try copying again!"
+      return "there is no valid email in the clipboard, try copying again!"
   }
   
   
@@ -309,3 +341,13 @@ function itNumber (clipData) {
 
     return clipData
 }
+
+
+
+// function hotKey (btn) {
+//     window.addEventListener("keydown",(e)=>{
+//         if (e.key === '`') {
+//             btn.click()
+//          }  
+//     })
+// }