+++ history.md
@@ -1,13 +1,15 @@
 ## Linkedin Registration Helper version history:
 ### 1.00 
-- Add UI 
-- Add automatic country selection by phone number
+- Added UI 
+- Added automatic country selection by phone number
 
 ### 1.01
 - Modification UI im sms-activate.org 
 
 ### 1.03
 - fix new UI updates im sms-activate.org 
-- add number highlighter
+- added number highlighter
+### 1.04
+- Added auto submit of forms after paste data